refactor of header
