+++ package.json
@@ -12,8 +12,8 @@
         "@chakra-ui/button": "^1.3.2",
         "@chakra-ui/icons": "^1.0.13",
         "@chakra-ui/layout": "^1.4.5",
-        "@chakra-ui/react": "^1.6.3",
         "@chakra-ui/radio": "^1.3.6",
+        "@chakra-ui/react": "^1.6.3",
         "@emotion/react": "^11.4.0",
         "@emotion/styled": "^11.3.0",
         "@stripe/stripe-js": "^1.13.2",
@@ -40,6 +40,7 @@
         "eslint-plugin-react": "^7.23.2",
         "eslint-plugin-react-hooks": "^4.2.0",
         "react-icons": "^4.2.0",
-        "typescript": "^4.2.4"
+        "typescript": "^4.2.4",
+        "utilities-techsweave": "^1.1.14"
     }
 }
