+++ src/components/header/MenuToggle.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { Box } from '@chakra-ui/layout';
+import { CloseIcon, HamburgerIcon } from '@chakra-ui/icons';
+
+type Props = {
+  toggle,
+  isOpen
+};
+const MenuToggle = ({ toggle, isOpen }:Props) => (
+  <Box display={{ base: 'block', md: 'none' }} onClick={toggle}>
+    {isOpen ? <CloseIcon /> : <HamburgerIcon />}
+  </Box>
+);
+export default MenuToggle;
